chore: 完成项目多处代码与配置更新
1. 修改jpa ddl-auto为update
2. 替换实体类id生成策略为@UuidGenerator配合@GeneratedValue
3. 修正订单表price字段类型为DECIMAL(10,2)
4. 调整ProductService分页查询逻辑
5. 更新用户角色绑定SQL脚本
6. 为Review添加评分校验注解
7. 初始化Order的items集合
8. 新增avatarUrl字段与Pageable依赖

+++ server/src/main/resources/db/migration/V1__create_initial_schema.sql
@@ -33,7 +33,7 @@ CREATE TABLE products (
 CREATE TABLE orders (
     id                  VARCHAR(36)    PRIMARY KEY,
     status              VARCHAR(15)    NOT NULL DEFAULT 'PENDING',
-    price               DOUBLE PRECISION NOT NULL,
+    price               DECIMAL(10,2) NOT NULL,
     payment_method      VARCHAR(25)    NOT NULL,
     address             TEXT           NOT NULL,
     delivery_method     VARCHAR(30)    NOT NULL,
